# code/static/swagger.yaml
openapi: 3.0.0
info:
  title: Chat API
  description: API specification for the Chat backend
  version: 1.0.0
servers:
  - url: https://staging.api.smalltech.in
    description: Staging server running on Google Cloud Platform (GCP)

components:
  schemas:
    status:
      type: string
      description: >
        Status of the lead.  
        Allowed values:  
        - OPEN → New lead, not yet processed
        - CLOSED → Lead is closed  
        - QUALIFYING → Lead is in process  
      enum: [OPEN, CLOSED, QUALIFYING]
      example: OPEN

paths:
  /health:
    get:
      summary: Health check endpoint
      description: Returns a hello world message to verify the service is up.
      responses:
        "200":
          description: Successful health check
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Hello World

  /chat:
    post:
      summary: Chat with the bot
      description: Send a user message and receive a chatbot response.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - session_id
                - input
              properties:
                input:
                  type: string
                  description: The user’s message
                  example: "Hello bot"
                session_id:
                  type: string
                  format: UUID
                  description: Session identifier for maintaining context. **Must be a valid UUID.**
                  example: "0b3cf7e1-5b30-46df-b018-85ca4dbd4391"
                request_type:
                  type: string
                  description: The type of the query/agent required
                  example: "sales"
                host:
                  type: string
                  description: website from which request in coming
                  example: "example.com"
                
      responses:
        "200":
          description: Successful response from the chatbot
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  response:
                    type: string
                    description: Bot response message
                    example: "Hi there! How can I help you today?"
        "400":
          description: Invalid input or address does not exist
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: string
                oneOf:
                  - properties:
                      error:
                        example: "Input cannot be empty."
                  - properties:
                      error:
                        example: "Enter correct address."

        "500":
          description: Server error during LLM call
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: string
                    example: "Sorry, something went wrong while processing your message. Please try again later."
  /prompts:
    get:
      summary: Get all prompts
      description: Returns all prompts from the prompts table.
      responses:
        "200":
          description: List of all prompts
          content:
            application/json:
              schema:
                type: object
                properties:
                  prompts:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                        domain:
                          type: string
                        agent_type:
                          type: string
                        type:
                          type: string
                        text:
                          type: string
                        created_at:
                          type: string
                          format: date-time
  /prompt:
    post:
      summary: Create or update a prompt
      description: Create a new prompt or update an existing one by (domain, agent_type, type).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - domain
                - agent_type
                - type
                - text
              properties:
                domain:
                  type: string
                  example: "common"
                agent_type:
                  type: string
                  example: "sales"
                type:
                  type: string
                  example: "intro-message"
                text:
                  type: string
                  example: "Welcome to our service!"
      responses:
        "200":
          description: Prompt created or updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Prompt created/updated."
        "400":
          description: Missing required fields
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: string
                    example: "Missing required fields: domain, agent_type, type, text"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: string
                    example: "Failed to create/update prompt."
  /history:
    get:
      summary: Get or create chat history
      description: >
        Returns the chat history for the given `session_id`.  
        If the session does not exist, a new one is created and initialized with the first AI message.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - session_id
              properties:
                session_id:
                  type: string
                  format: UUID
                  description: Unique session identifier of the lead
                  example: "0b3cf7e1-5b30-46df-b018-85ca4dbd4391"
                host:
                  type: string
                  description: website from which request in coming
            examples:
              withHost:
                summary: contains both host and session id
                value:
                  session_id: "0b3cf7e1-5b30-46df-b018-85ca4dbd4391"
                  host: "example.com"
              withoutHost:
                summary: No need to send host explictly, and host is taken from request header
                value:
                  session_id: "0b3cf7e1-5b30-46df-b018-85ca4dbd4391"
      responses:
        '200':
          description: Chat history retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  history:
                    type: array
                    items:
                      type: object
                      properties:
                        type:
                          type: string
                          enum: [human, ai]
                          description: Sender of the message
                        content:
                          type: string
                          description: Message text
                example:
                  history:
                    - type: human
                      content: "Hello"
                    - type: ai
                      content: "Hi there! How can I help you?"
        '201':
          description: New chat session created with first AI message
          content:
            application/json:
              schema:
                type: object
                properties:
                  session_id:
                    type: string
                    format: UUID
                    description: Newly created session identifier
                  history:
                    type: array
                    items:
                      type: object
                      properties:
                        type:
                          type: string
                          enum: [human, ai]
                          description: Sender of the message
                        content:
                          type: string
                          description: Message text
                example:
                  session_id: "xyz789"
                  history:
                    - type: ai
                      content: "Hello! I’m your assistant. How can I help you today?"
        '400':
          description: Invalid session_id format or incorrect address
        '500':
          description: Server error
  /chat-info:
    get:
      summary: Retrieve stored chat info
      description: >
        Fetch all chat info records stored in the database.  
        Each record contains session details such as name, email, and mobile number.
      responses:
        "200":
          description: Successfully retrieved chat info
          content:
            application/json:
              schema:
                type: object
                properties:
                  chat_info:
                    type: array
                    items:
                      type: object
                      properties:
                        session_id:
                          type: string
                          format: UUID
                          description: Session identifier
                          example: "0b3cf7e1-5b30-46df-b018-85ca4dbd4391"
                        name:
                          type: string
                          description: Name of the user
                          example: "Vivek Agarwal"
                        email:
                          type: string
                          format: email
                          description: User's email address
                          example: "vivek@example.com"
                        mobile_number:
                          type: string
                          description: User's mobile phone number
                          example: "+91-9876543210"
                        country:
                          type: string
                          description: User's country
                          example: "India"
                        status:
                          type: string
                          enum: [OPEN, CLOSED, QUALIFYING]
                          description: Analyst status for that session id
                          example: "OPEN"

                        remarks:
                          type: string
                          description: Analyst's remark for that session id
                          example: "Send a mail and waiting for a review"

                        doamin:
                          type: string
                          description: request frontend's Domain
                          example: "SMALLTECH"

                        time:
                          type: string
                          description: Date and time when contact info is detected
                          example: "2025-12-01 12:38:54.331648+05:30"

        "500":
          description: Server error while fetching chat info
          content:
            application/json:
              schema:
                type: object
                properties:

                  error:
                    type: string
                    example: "Unable to fetch chat info. Please try again later."
                    
    patch:
      summary: Update lead status, remarks and is_active
      description: >
        Update the `status` and/or `remarks` and/or `is_active` of a lead identified by its `session_id`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - session_id
              properties:
                session_id:
                  type: string
                  format: UUID
                  description: Unique session identifier of the lead
                  example: "0b3cf7e1-5b30-46df-b018-85ca4dbd4391"
                status:
                  $ref: '#/components/schemas/status'
                remarks:
                  type: string
                  description: Analyst's updated remarks
                is_active:
                  type: boolean
                  description: Used for soft delete unneccessary data
            examples:
              updateStatusOnly:
                summary: Update only status
                value:
                  session_id: "0b3cf7e1-5b30-46df-b018-85ca4dbd4391"
                  status: "CLOSED"
              updateRemarksOnly:
                summary: Update only remarks
                value:
                  session_id: "0b3cf7e1-5b30-46df-b018-85ca4dbd4391"
                  remarks: "Sent follow-up email, awaiting response"
              updateis_activeOnly:
                summary: Update only is_active
                value:
                  session_id: "0b3cf7e1-5b30-46df-b018-85ca4dbd4391"
                  is_active: false
              updateStatusAndRemarks:
                summary: Update Any two (status and remarks)
                value:
                  session_id: "0b3cf7e1-5b30-46df-b018-85ca4dbd4391"
                  status: "QUALIFYING"
                  remarks: "Client called back, demo scheduled"
              updateStatusAndRemarksAndis_active:
                summary: Update All (status, remarks and is_active)
                value:
                  session_id: "0b3cf7e1-5b30-46df-b018-85ca4dbd4391"
                  status: "QUALIFYING"
                  remarks: "Client called back, demo scheduled"
                  is_active: false
      responses:
        "200":
          description: Chat-info updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Chat-info updated successfully."
                  updated_lead:
                    type: object
                    properties:
                      session_id:
                        type: string
                        format: UUID
                        example: "0b3cf7e1-5b30-46df-b018-85ca4dbd4391"
                      status:
                        $ref: '#/components/schemas/status'
                      remarks:
                        type: string
                        example: "Client called back, demo scheduled"
                      is_active:
                        type: boolean
                        example: false
        "400":
          description: Invalid request payload
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: string
                    example: "Enter correct input"
        "500":
          description: Server error while updating lead
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: string
                    example: "Unable to update lead. Please try again later."

  /chat-info/contact:
    patch:
      summary: Update contact info for a session
      description: >
        Update the contact details (`name`, `email`, `mobile`, `country`) of a lead identified by its `session_id`.  
        All contact fields are optional, but at least one must be provided.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - session_id
              properties:
                session_id:
                  type: string
                  format: UUID
                  description: Unique session identifier of the lead
                  example: "0b3cf7e1-5b30-46df-b018-85ca4dbd4391"
                name:
                  type: string
                  description: Full name of the contact
                  example: "Jane Doe"
                email:
                  type: string
                  format: email
                  description: Email address of the contact
                  example: "jane@example.com"
                mobile:
                  type: string
                  description: Mobile phone number of the contact
                  example: "+919876543210"
                country:
                  type: string
                  description: Country of the contact
                  example: "India"
            examples:
              updateNameOnly:
                summary: Update only name
                value:
                  session_id: "0b3cf7e1-5b30-46df-b018-85ca4dbd4391"
                  name: "Jane Doe"
              updateEmailOnly:
                summary: Update only email
                value:
                  session_id: "0b3cf7e1-5b30-46df-b018-85ca4dbd4391"
                  email: "jane@example.com"
              updateAll:
                summary: Update all contact fields
                value:
                  session_id: "0b3cf7e1-5b30-46df-b018-85ca4dbd4391"
                  name: "Jane Doe"
                  email: "jane@example.com"
                  mobile: "+919876543210"
                  country: "India"
      responses:
        "200":
          description: Contact info updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "contact info updated"
        "400":
          description: Invalid request payload
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: string
                    examples:
                      missingFields:
                        value: "At least one of name, email, mobile, or country is required"
                      invalidEmail:
                        value: "Invalid email address"
                      invalidMobile:
                        value: "Invalid mobile number"
                      invalidSession:
                        value: "Invalid session id format"
        "500":
          description: Server error while updating contact info
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: string
                    example: "Sorry, something went wrong. Please try again later."